move remove_cryosparc_leading_uid() to metadata.py as originally requested
